ch&jm - Refactored collision framework
Collisions now handled in entities and colliders for the most part. Needs
some testing still.
