+++ Server/GameServer.cpp
@@ -9,19 +9,6 @@
 #include "SBoxEntity.hpp"
 #include "GameServer.hpp"
 
-// Allow pairs inside unordered_set
-struct PairHash
-{
-	template <class T1, class T2>
-	std::size_t operator () (std::pair<T1, T2> const &pair) const
-	{
-		std::size_t h1 = std::hash<T1>()(pair.first);
-		std::size_t h2 = std::hash<T2>()(pair.second);
-
-		return h1 ^ h2;
-	}
-};
-
 GameServer::GameServer()
 {
 }
@@ -51,12 +38,17 @@ void GameServer::start()
 	// Init level parser and load level
 	_levelParser = std::make_unique<GridLevelParser>();
 
+	// Init collision manager
+	_collisionManager = std::make_unique<CollisionManager>(&_entityMap, &_jails);
+
 	// Map initialization
 	auto entities = _levelParser->parseLevelFromFile(
 		"Levels/map.dat",
 		_jails,
 		_humanSpawns,
 		_dogSpawns);
+
+	Logger::getInstance()->debug("Parsed " + std::to_string(entities.size()) + " entities from file.");
 
 	// Ensure at least one human spawn, dog spawn, and jail
 	// TODO: convert these to exceptions, or find a cleaner way to handle this
@@ -131,7 +123,7 @@ void GameServer::update()
 					_dogSpawns.push(dogSpawn);
 
 					// Create player entity at position
-					playerEntity = std::make_shared<SDogEntity>(playerEvent->playerId);
+					playerEntity = std::make_shared<SDogEntity>(playerEvent->playerId, &_jails);
 					playerEntity->getState()->pos = glm::vec3(dogSpawn.x, 0, dogSpawn.y);
 					_dogs.push_back(playerEntity);
 				}
@@ -199,7 +191,7 @@ void GameServer::update()
 		}
 
 		// Collision resolution
-		handleCollisions();
+		_collisionManager->handleCollisions();
 
 		// Update state of the game
 		bool dogsCaught = true;
@@ -266,287 +258,4 @@ void GameServer::update()
 			_gameDuration += (std::chrono::steady_clock::now() - timerStart);
 		}
     }
-}
-
-void GameServer::handleCollisions()
-{
-	// Build quadtree
-	auto tree = new QuadTree({ glm::vec2(0), MAP_WIDTH / 2 });
-	for (auto& entityPair : _entityMap)
-	{
-		tree->insert(entityPair.second->getState().get());
-	}
-
-	auto collisionSet = std::unordered_set<std::pair<BaseState*, BaseState*>, PairHash>();
-
-	// Build set of pairs of collisions
-	for (auto& entityPair : _entityMap)
-	{
-		auto entity = entityPair.second;
-
-		// Only run collision check if not static
-		if (!entity->getState()->isStatic)
-		{
-			for (auto& collidingEntity : entity->getColliding(*tree))
-			{
-				collisionSet.insert({ entity->getState().get(), collidingEntity });
-			}
-		}
-	}
-
-	// Iterate until no collisions
-	while (!collisionSet.empty())
-	{
-		auto collisionPair = (collisionSet.begin());
-		auto objectA = collisionPair->first;
-		auto objectB = collisionPair->second;
-
-		// Erase from beginning
-		collisionSet.erase(collisionPair);
-
-		if (objectA->type == ENTITY_DOG && objectB->type == ENTITY_HUMAN)
-		{
-			handleDogCaught(objectA, objectB, collisionSet, tree);
-			continue;
-		}
-
-		if (objectA->type == ENTITY_HUMAN && objectB->type == ENTITY_DOG)
-		{
-			handleDogCaught(objectB, objectA, collisionSet, tree);
-			continue;
-		}
-
-		// We know that A is a player (capsule collider), so switch on b only
-		switch (objectB->colliderType)
-		{
-		case COLLIDER_CAPSULE:
-		{
-			handleCapsule(objectA, objectB);
-			break;
-		}
-
-		case COLLIDER_AABB:
-		{
-			handleAABB(objectA, objectB);
-			break;
-		}
-
-		} // switch
-
-		// Remove duplicates (e.g. <A,B> vs <B,A> if both players)
-		if (!objectB->isStatic)
-		{
-			collisionSet.erase({ objectB, objectA });
-
-			// Mark as changed
-			auto result = _entityMap.find(objectB->id);
-
-			if (result != _entityMap.end())
-			{
-				result->second->hasChanged = true;
-			}
-		}
-
-		// Find entity for object A
-		auto result = _entityMap.find(objectA->id);
-
-		if (result == _entityMap.end())
-		{
-			throw std::runtime_error("Could not find entity in handleCollisions()");
-		}
-
-		auto entity = result->second;
-		entity->hasChanged = true;
-
-		// Re-check for colliding
-		for (auto& collidingEntity : entity->getColliding(*tree))
-		{
-			collisionSet.insert({ entity->getState().get(), collidingEntity });
-		}
-	}
-
-	delete tree;
-}
-
-
-void GameServer::handleAABB(BaseState* stateA, BaseState* stateB)
-{
-	float rA = (float)std::fmax(stateA->width, stateA->depth) / 2;
-
-	// Check which side of box the collision happens on
-	float dists[4];
-	dists[0] = (stateB->pos.x - stateB->width / 2) - stateA->pos.x; // West
-	dists[1] = stateA->pos.x - (stateB->pos.x + stateB->width / 2); // East
-	dists[2] = stateA->pos.z - (stateB->pos.z + stateB->depth / 2); // North
-	dists[3] = (stateB->pos.z - stateB->depth / 2) - stateA->pos.z; // South
-
-	// Check if collision happens at corner
-	glm::vec3 cornerPos = stateB->pos;
-	bool inCorner = false;
-	if (dists[0] > 0) {
-		cornerPos.x -= stateB->width / 2;
-		if (dists[2] > 0) {
-			cornerPos.z += stateB->depth / 2;
-			inCorner = true;
-		}
-		else if (dists[3] > 0) {
-			cornerPos.z -= stateB->depth / 2;
-			inCorner = true;
-		}
-	} else if (dists[1] > 0) {
-		cornerPos.x += stateB->width / 2;
-		if (dists[2] > 0) {
-			cornerPos.z += stateB->depth / 2;
-			inCorner = true;
-		}
-		else if (dists[3] > 0) {
-			cornerPos.z -= stateB->depth / 2;
-			inCorner = true;
-		}
-	}
-
-	// Collision happens at corner, 
-	if (inCorner) {
-		// Vector from corner to A
-		glm::vec3 diff = stateA->pos - cornerPos;
-		
-		// We don't considerate height
-		diff.y = 0;
-
-		float overlap = rA - glm::length(diff);
-
-		// Vector to move circles by
-		glm::vec3 correctionVec;
-
-		// Normal case: some displacement between circles
-		if (glm::length(diff))
-		{
-			// How much the circles overlap, and the ratio of overlap to distance
-			// between circles
-			float ratio = overlap / glm::length(diff);
-			correctionVec = diff * ratio;
-		}
-		// Edge case: objects directly on top each other
-		else
-		{
-			correctionVec = glm::vec3(1, 0, 0) * overlap;
-		}
-
-		// Apply to A
-		stateA->pos += correctionVec;
-	}
-	else {
-		int minIndex = -1;
-		float min = FLT_MAX;
-
-		// Get closest edge
-		for (int i = 0; i < 4; i++)
-		{
-			if (dists[i] > 0)
-			{
-				minIndex = i;
-				break;
-			}
-			
-			if (dists[i] < min)
-			{
-				min = dists[i];
-				minIndex = i;
-			}
-		}
-
-		glm::vec3 correctionVec = glm::vec3(0);
-		switch (minIndex)
-		{
-		case 0: // West
-			correctionVec.x = -(rA - dists[0]);
-			break;
-		case 1: // East
-			correctionVec.x = rA - dists[1];
-			break;
-		case 2: // North
-			correctionVec.z = rA - dists[2];
-			break;
-		case 3: // South
-			correctionVec.z = -(rA - dists[3]);
-			break;
-		}
-
-		stateA->pos += correctionVec;
-	}
-}
-
-
-void GameServer::handleCapsule(BaseState* stateA, BaseState* stateB)
-{
-	float rA = (float)std::fmax(stateA->width, stateA->depth) / 2;
-	float rB = (float)std::fmax(stateB->width, stateB->depth) / 2;
-
-	// Vector from B to A
-	glm::vec3 diff = stateA->pos - stateB->pos;
-
-	float overlap = rA + rB - glm::length(diff);
-
-	// Vector to move circles by
-	glm::vec3 correctionVec;
-
-	// Normal case: some displacement between circles
-	if (glm::length(diff))
-	{
-		// How much the circles overlap, and the ratio of overlap to distance
-		// between circles
-		float ratio = overlap / glm::length(diff);
-		correctionVec = diff * ratio;
-	}
-	// Edge case: objects directly on top each other
-	else
-	{
-		correctionVec = glm::vec3(1, 0, 0) * overlap;
-	}
-
-	// If two movable objects, only move halfway
-	if (!stateB->isStatic)
-	{
-		correctionVec /= 2.0f;
-
-		// Apply to B
-		stateB->pos -= correctionVec;
-	}
-
-	// Apply to A
-	stateA->pos += correctionVec;
-}
-
-void GameServer::handleDogCaught(
-	BaseState* dog,
-	BaseState* human,
-	std::unordered_set<std::pair<BaseState*, BaseState*>, PairHash> & collisionSet,
-	QuadTree * tree)
-{
-	// Find entity for dog
-	auto result = _entityMap.find(dog->id);
-
-	if (result == _entityMap.end())
-	{
-		std::runtime_error("Could not find entity in handleCollisions()");
-	}
-
-	auto entity = result->second;
-	entity->hasChanged = true;
-
-	// Cast as dog entity and mark as caught
-	auto dogEntity = std::static_pointer_cast<SDogEntity>(entity);
-	dogEntity->isCaught = true;
-
-	// Removing all colliding with dog
-	for (auto& collidingEntity : entity->getColliding(*tree))
-	{
-		collisionSet.erase({ entity->getState().get(), collidingEntity });
-		collisionSet.erase({ collidingEntity, entity->getState().get() });
-	}
-
-	// Get a random jail location
-	std::random_shuffle(_jails.begin(), _jails.end());
-	glm::vec2 jailPos = _jails[0];
-	dog->pos = glm::vec3(jailPos.x, 0, jailPos.y);
 }
